requirements.txt
PySide6>=6.5
pyirsdk>=0.2
requests>=2.28